+++ test/model/tetrimino/type/ttetrimino_test.cpp
@@ -58,8 +58,8 @@ TEST_CASE("class TTetrimino") {
     REQUIRE(tetrimino.X() == 3);
     REQUIRE(tetrimino.Y() == 0);
 
-    for (int line = 0; line < tetrimino.minos().size(); ++line) {
-      for (int column = 0;
+    for (size_t line = 0; line < tetrimino.minos().size(); ++line) {
+      for (size_t column = 0;
            column < tetrimino.minos().at(tetrimino.orientation())[line].size();
            ++column) {
         REQUIRE(tetrimino.minos().at(tetrimino.orientation())[line][column] ==
@@ -72,8 +72,8 @@ TEST_CASE("class TTetrimino") {
       auto tetrimino = TTetrimino();
       tetrimino.rotate(true);
 
-      for (int line = 0; line < tetrimino.minos().size(); ++line) {
-        for (int column = 0;
+      for (size_t line = 0; line < tetrimino.minos().size(); ++line) {
+        for (size_t column = 0;
              column <
              tetrimino.minos().at(tetrimino.orientation())[line].size();
              ++column) {
@@ -87,8 +87,8 @@ TEST_CASE("class TTetrimino") {
       tetrimino.rotate(true);
       tetrimino.rotate(true);
 
-      for (int line = 0; line < tetrimino.minos().size(); ++line) {
-        for (int column = 0;
+      for (size_t line = 0; line < tetrimino.minos().size(); ++line) {
+        for (size_t column = 0;
              column <
              tetrimino.minos().at(tetrimino.orientation())[line].size();
              ++column) {
@@ -101,8 +101,8 @@ TEST_CASE("class TTetrimino") {
       auto tetrimino = TTetrimino();
       tetrimino.rotate(false);
 
-      for (int line = 0; line < tetrimino.minos().size(); ++line) {
-        for (int column = 0;
+      for (size_t line = 0; line < tetrimino.minos().size(); ++line) {
+        for (size_t column = 0;
              column <
              tetrimino.minos().at(tetrimino.orientation())[line].size();
              ++column) {
